Add support for python 3.14

See merge request jatic/aria/dataeval!958

+++ src/dataeval/core/_clusterer.py
@@ -202,7 +202,7 @@ class _KMeansSorter:
             np.max(n_samples_per_cluster),
             np.mean(n_samples_per_cluster),
         )
-        return _Clusters(clst.labels_, clst.cluster_centers_)
+        return _Clusters(clst.labels_.astype(np.intp), clst.cluster_centers_.astype(np.float64))
 
 
 class _HDBSCANSorter:
